refactor: Badge 타입과 import 정리
