+++ packages/jds/src/components/Badge/dotBadge/dotBadge.types.ts
@@ -1,14 +1,7 @@
-import type { FeedbackVariant, BadgeSize } from "./dotBadge.types";
-
-export {
-  BADGE_SIZE_OPTIONS,
-  FEEDBACK_VARIANT_OPTIONS,
-  type BadgeSize,
-  type FeedbackVariant,
-} from "../badge.types";
+import type { BadgeSize, FeedbackVariant } from "../badge.types";
 
 export interface DotBadgeFeedbackProps {
-  variant: FeedbackVariant;
+  variant?: FeedbackVariant;
   size?: BadgeSize;
   isMuted?: boolean;
 }
